Adding guard to avoid "already closed" exception from MapDB
If shutdown/undeploy of the repository and subsequently the index provider is called from a JVM shutdown hook, e.g. in the case of a spring managed scenario, then it is possible that the shutdown hook registered via MapDB will be call prior to the call to postShutdown.  In this scenario, MapDB will throw an IllegalStateException "already closed".  Adding simple guard to check for whether the MapDb instance is already closed.